backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
sqlalchemy-utils
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-multipart
pydantic[email]